update diags - not fully tested yet
git-svn-id: https://svn-ccsm-models.cgd.ucar.edu/postprocessing/trunk@75669 fe37f545-8307-0410-aea5-b40df96820b5

+++ diagnostics/diagnostics/ocn/Plots/ocn_diags_plot_bc.py
@@ -82,22 +82,24 @@ class OceanDiagnosticPlot(object):
 
                 sourceFile = '{0}/{1}.ps'.format(workdir, psFile)
 ##                print('...... convert source file {0}'.format(sourceFile))
+                # check if the sourceFile exists
+                rc, err_msg = cesmEnvLib.checkFile(sourceFile,'read')
 
-                # check if the image file alreay exists and remove it to regen
+                # check if the image file already exists and remove it to regen
                 imgFile = '{0}/{1}.{2}'.format(workdir, psFile, imgFormat)
-                rc, err_msg = cesmEnvLib.checkFile(imgFile,'write')
-                if rc:
+                rc1, err_msg1 = cesmEnvLib.checkFile(imgFile,'write')
+                if rc and rc1:
                     print('...... removing {0} before recreating'.format(imgFile))
                     os.remove(imgFile)
 
-                # convert the image from ps to imgFormat
-                try:
-                    pipe = subprocess.check_call( ['convert', '-trim', '-bordercolor', 'white', '-border', '5x5', '-density', '95', '{0}'.format(sourceFile),'{0}'.format(imgFile)])
+                    # convert the image from ps to imgFormat
+                    try:
+                        pipe = subprocess.check_call( ['convert', '-trim', '-bordercolor', 'white', '-border', '5x5', '-density', '95', '{0}'.format(sourceFile),'{0}'.format(imgFile)])
 ##                   print('...... created {0} size = {1}'.format(imgFile, os.path.getsize(imgFile)))
-                except subprocess.CalledProcessError as e:
-                    print('...... failed to create {0}'.format(imgFile))
-                    print('WARNING: convert_plots call to convert failed with error:')
-                    print('    {0}'.format(e.output))
+                    except subprocess.CalledProcessError as e:
+                        print('...... failed to create {0}'.format(imgFile))
+                        print('WARNING: convert_plots call to convert failed with error:')
+                        print('    {0}'.format(e.output))
                 else:
                     continue
         else:
